chore: move error middleware to controllers/
